feat: Generate & apply JWT for valid user
