--- src/main/java/com/tringles/tutorial/config/OAuth2SuccessHandler.java
@@ -1,40 +1,0 @@
-import com.tringles.tutorial.domain.User;
-import com.tringles.tutorial.service.UserService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.oauth2.core.oidc.user.OidcUser;
-import org.springframework.security.oauth2.core.user.DefaultOAuth2User;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
-import org.springframework.stereotype.Component;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@Component
-public class OAuth2SuccessHandler implements AuthenticationSuccessHandler {
-
-    @Autowired
-    private UserService userService;
-
-    @Override
-    public void onAuthenticationSuccess(
-            HttpServletRequest request,
-            HttpServletResponse response,
-            Authentication authentication) throws IOException, ServletException {
-        Object principal = authentication.getPrincipal();
-        if (principal instanceof DefaultOAuth2User) {
-            com.tringles.tutorial.domain.OAuth2User oauth = com.tringles.tutorial.domain.OAuth2User
-                    .Provider.google.convert((OAuth2User) principal);
-            User user = userService.load(oauth);
-            SecurityContextHolder.getContext().setAuthentication(
-                    new UsernamePasswordAuthenticationToken(user, null, user.getAuthorities())
-            );
-        }
-        request.getRequestDispatcher("/").forward(request, response);
-    }
-}
